backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.9.2
tensorflow==2.16.2
numpy==1.26.4
pillow==10.4.0
scikit-learn==1.5.2
kaggle==1.6.17
